lambda_functions/document_processor/requirements.txt
boto3
openai
chromadb==0.4.22
numpy==1.26.4
pysqlite3-binary
python-docx
PyPDF2
pypandoc
markdown
langchain
langchain-openai
tiktoken